requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cryptography==45.0.2
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
ExifRead==3.3.1
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Markdown==0.3
Flask-PyMongo==3.0.1
Flask-WTF==1.2.2
frozenlist==1.6.0
gunicorn==23.0.0
h11==0.16.0
http_ece==1.2.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.2.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.4
numpy==2.2.6
ordered-set==4.1.0
packaging==25.0
pillow==11.2.1
pillow_heif==0.22.0
pip-review==1.3.0
propcache==0.3.1
py-vapid==1.9.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.13.0
pypng==0.20220715.0
python-dotenv==1.1.0
pywebpush==2.0.3
qrcode==8.2
requests==2.32.3
rich==13.9.4
segno==1.6.6
segno-pil==1.0.0
sinch==1.1.1
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
vonage==4.4.3
vonage-account==1.1.1
vonage-application==2.0.1
vonage-http-client==1.5.1
vonage-jwt==1.1.5
vonage-messages==1.4.0
vonage-network-auth==1.0.2
vonage-network-number-verification==1.0.2
vonage-network-sim-swap==1.1.2
vonage-number-insight==1.0.7
vonage-numbers==1.0.4
vonage-sms==1.1.6
vonage-subaccounts==1.0.4
vonage-users==1.2.1
vonage-utils==1.1.4
vonage-verify==2.1.0
vonage-verify-legacy==1.0.1
vonage-video==1.2.0
vonage-voice==1.4.0
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
yarl==1.20.0
zipp==3.21.0
